requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
